internal/hivedocker: rename to libdocker
