+++ internal/libdocker/container.go
@@ -1,4 +1,4 @@
-package hivedocker
+package libdocker
 
 import (
 	"archive/tar"
